requirements.txt
spacy
pdfminer.six
python-docx
scikit-learn
pandas
numpy
pytest
pytest==7.4.3
transformers
torch
sentence-transformers
dataset
fpdf2
fastapi
uvicorn
python-multipart
datasets

